requirements.txt
accelerate
bitsandbytes
datasets
einops
isort
jsonlines
loralib
optimum
packaging
peft
ray
tensorboard
torchmetrics
tqdm
transformers==4.46.3
transformers_stream_generator
wandb
wheel
